refactor(#278): extract GroundingMixin from query.py
PR #276's review flagged query.py at 1126 lines as crossing the project's
review threshold, with the grounding cluster forming a natural seam:
polarity-axis caching + grounding strength calculation have no dependency
on the rest of QueryMixin (vector search, learned knowledge), and the
rest of QueryMixin has no dependency on them. Splitting here lets future
grounding changes land without touching unrelated query paths.

## What moved

To new api/app/lib/age_client/grounding.py as GroundingMixin:
- Module docstring describing the two-tier cache architecture
- Module-level cache globals + locks (_polarity_axis_cache,
  _grounding_cache, _grounding_cache_generation)
- _parse_embedding, _parse_confidence helpers (used only by the
  grounding code in this package)
- _get_vocab_generation, _get_polarity_axis methods
- calculate_grounding_strength_semantic, calculate_grounding_strength_batch
- BATCH_CHUNK_SIZE import (only used by the batch method)

query.py drops from 1141 lines to 604 — back under the project's
priority band. The remaining content is what the file's name actually
describes: vector search, learned knowledge CRUD, and the three lazy
facade properties (facade, graph, epochs).

## Composition wiring

api/app/lib/age_client/__init__.py adds GroundingMixin to AGEClient's
MRO chain, slotted between OntologyMixin and QueryMixin (alphabetical
within the cluster of "data-shape mixins"). The import surface for
callers is unchanged — `from api.app.lib.age_client import AGEClient`
keeps working.

## Test updates

tests/unit/lib/test_batch_chunk_recovery.py (from commit 3) updates its
imports from `query` to `grounding` — same test bodies, just pointed at
the new module. The fakes and assertions are unchanged.

## Why not split into smaller PRs

The warm-cache short-circuit and the cache-invariant docstrings also
target this same file pair. Doing the file move alone first lets each
follow-up commit land as a small, focused change against the new
GroundingMixin instead of editing a moving target. See the next two
commits.

Full suite: 1079 passed, 15 skipped.

Closes #278 (file extraction portion; warm-cache + docstrings follow
in subsequent commits).
